[branch] The version of my Cashe increase. Now it is becoming testing. Correct mistakes in Cashe::Cashe and Cashe::processRead.

+++ branches/2011/sandbox/devlanov_cashe/cashe.h
@@ -22,9 +22,9 @@ private:
     unsigned int ways;
     unsigned int block_size;
     unsigned int addr_size_in_bit;
-    unsigned int miss_ratio_priv;
-    unsigned int hit_ratio_priv;
-    unsigned int number_all_request_priv;
+    long double miss_ratio_priv;
+    long double hit_ratio_priv;
+    long double number_all_request_priv;
     unsigned int number_sets_priv;
     unsigned int bit_in_cashe_index; 
     StructItem** cash_array;
